Reorganize entries folder structure, minor styling changes
